Add new dynamic versions table
This makes use of JS and the new Parchment mappings version endpoint to
dynamically list out the latest version for each available Minecraft
version.

The current implementation sorts the MC versions using the info from the
launcher manifest, and falling back to a hardcoded array if that is not
available.

The current impl. also replaces the rows of the table with new rows,
so that the original rows (which uses version badges) of the table remain
if the script cannot access the versions endpoint.